% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNarrData-package.R
\docType{package}
\name{addNarrData-package}
\alias{addNarrData}
\alias{addNarrData-package}
\title{addNarrData: Adds NARR Variables Based on Date and 12 x 12 km Grid Cell}
\description{
Adds NARR weather variables based on date and 12 x 12 km grid cell.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://geomarker.io/add_narr_data}
  \item Report bugs at \url{https://github.com/geomarker-io/add_narr_data/issues}
}

}
\author{
\strong{Maintainer}: Erika Rasnick \email{erika.rasnick@cchmc.org}

Authors:
\itemize{
  \item Cole Brokamp
}

}
\keyword{internal}
