% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narr_data_functions.R
\name{download_narr_fst}
\alias{download_narr_fst}
\title{download narr.fst file to application specific directory so that it can be shared across R sessions and projects}
\usage{
download_narr_fst()
}
\description{
download narr.fst file to application specific directory so that it can be shared across R sessions and projects
}
