% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narr_cell_functions.R
\name{get_narr_cell_numbers}
\alias{get_narr_cell_numbers}
\title{get NARR cell numbers for given lat and lon}
\usage{
get_narr_cell_numbers(d)
}
\arguments{
\item{d}{data.frame with columns 'lat' and 'lon'}
}
\value{
a data.frame identical to the input data.frame but with appended NARR cell numbers
}
\description{
get NARR cell numbers for given lat and lon
}
\examples{
if (FALSE) {
d <- data.frame(
  id = c('1a', '2b', '3c'),
  lat = c(39.19674, 39.19674, 39.48765),
  lon = c(-84.582601, -84.582601, -84.610173)
)

get_narr_cell_numbers(d)
}
}
