% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narr_data_functions.R
\name{narr_fst}
\alias{narr_fst}
\title{checks for and returns filepath to narr.fst file in application-specific site data directory or current working directory
if not found, fails with suggestion to run download_narr_fst()}
\usage{
narr_fst()
}
\description{
checks for and returns filepath to narr.fst file in application-specific site data directory or current working directory
if not found, fails with suggestion to run download_narr_fst()
}
